dev 2024-12

+++ src/client/lib/software_install.ts
@@ -3,6 +3,16 @@ import { GlobalContextType } from "../providers/global.provider";
 import { fetchHtml } from "./utils.client";
 
 
+export type InstallMinerOptions = {
+    confirmed?: boolean,
+    // TODO
+}
+
+export type UninstallMinerOptions = {
+    confirmed?: boolean,
+}
+
+
 // @ts-ignore
 const alertify = window.alertify;
 
@@ -11,14 +21,14 @@ export const showUninstallMiner = (context: GlobalContextType, minerName: string
     if (! context.rigHost) return;
     console.log(`showUninstallMiner ${minerName}`)
 
-    // TODO
-    const minerAlias = `${minerName}-test-todo`; // TODO: proposer tous les alias installés pour ce miner
-    uninstallMiner(context, minerName, minerAlias);
+    // TODO ?
+    //const minerAlias = `${minerName}-test-todo`; // TODO: proposer tous les alias installés pour ce miner
+    //uninstallMiner(context, minerName, minerAlias);
 };
 
 
 
-export const installMiner = (context: GlobalContextType, minerName: string, minerAlias?: string, options?: {[key: string]: any}) => {
+export const installMiner = (context: GlobalContextType, minerName: string, minerAlias?: string, options?: InstallMinerOptions) => {
     if (! context.rigHost) return;
     console.log(`installMiner ${minerName} / ${minerAlias}`);
 
@@ -26,6 +36,7 @@ export const installMiner = (context: GlobalContextType, minerName: string, mine
     const onSuccess = (minerName: string, minerAlias: string | undefined, result: any) => {};
     const onFail = (minerName: string, minerAlias: string | undefined, err: any) => {};
 
+    const confirmed = options?.confirmed || false;
     let error: string | null = null;
 
     if (! minerName) {
@@ -41,54 +52,61 @@ export const installMiner = (context: GlobalContextType, minerName: string, mine
     //const minerFullName = `${minerName}-${minerAlias}`;
     const minerFullTitle = (minerName === minerAlias || ! minerAlias) ? minerName : `${minerName} (${minerAlias})`;
 
-    alertify.confirm("<b>Miner installation - confirmation</b>", `Do you want to install the miner '<b>${minerFullTitle}</b>' ?`,
-        function(){
-            alertify.success(`Starting miner ${minerFullTitle} installation...`);
+    function _installMiner() {
+        alertify.success(`Starting miner ${minerFullTitle} installation...`);
 
-            if (typeof onStart === 'function') {
-                onStart(minerName, minerAlias);
-            }
-
-            const data: {[key: string]: any} = {
-                action: 'start',
-                miner: minerName,
-                minerAlias,
-            };
-
-            const url = `http://${context.rigHost}/rig/miners/${minerName}/install`;
-
-            fetchHtml(url, { useProxy: true, method: 'POST', body: JSON.stringify(data) })
-                .then(({data, headers, status}) => {
-                    if (data && data.startsWith('OK:')) {
-                        if (typeof onSuccess === 'function') {
-                            onSuccess(minerName, minerAlias, data);
-                        }
-                        alertify.success(`Miner ${minerFullTitle} installation started`);
-
-                    } else {
-                        if (typeof onFail === 'function') {
-                            onFail(minerName, minerAlias, { message: data });
-                        }
-                        alertify.error(`Miner ${minerFullTitle} installation cannot be started. ${data}`);
-                    }
-
-                })
-                .catch((err: any) => {
-                    if (typeof onFail === 'function') {
-                        onFail(minerName, minerAlias, err);
-                    }
-
-                    alertify.error(`Miner ${minerFullTitle} installation cannot be started. ${err.message}`);
-                });
-        },
-        function(){
-            //alertify.error('Cancel');
+        if (typeof onStart === 'function') {
+            onStart(minerName, minerAlias);
         }
-    );
+
+        const data: {[key: string]: any} = {
+            action: 'start',
+            miner: minerName,
+            minerAlias,
+        };
+
+        const url = `http://${context.rigHost}/rig/miners/${minerName}/install`;
+
+        fetchHtml(url, { useProxy: true, method: 'POST', body: JSON.stringify(data) })
+            .then(({data, headers, status}) => {
+                if (data && data.startsWith('OK:')) {
+                    if (typeof onSuccess === 'function') {
+                        onSuccess(minerName, minerAlias, data);
+                    }
+
+                    alertify.success(`Miner ${minerFullTitle} installation started`);
+
+                } else {
+                    if (typeof onFail === 'function') {
+                        onFail(minerName, minerAlias, { message: data });
+                    }
+
+                    alertify.error(`Miner ${minerFullTitle} installation cannot be started. ${data}`);
+                }
+
+            })
+            .catch((err: any) => {
+                if (typeof onFail === 'function') {
+                    onFail(minerName, minerAlias, err);
+                }
+
+                alertify.error(`Miner ${minerFullTitle} installation cannot be started. ${err.message}`);
+            });
+    }
+
+    if (confirmed) {
+        _installMiner();
+
+    } else {
+        alertify.confirm("<b>Miner installation - confirmation</b>", `Do you want to install the miner '<b>${minerFullTitle}</b>' ?`,
+            _installMiner,
+            () => { /* alertify.error('Cancel'); */ }
+        );
+    }
 };
 
 
-export const uninstallMiner = (context: GlobalContextType, minerName: string, minerAlias?: string) => {
+export const uninstallMiner = (context: GlobalContextType, minerName: string, minerAlias?: string, options?: UninstallMinerOptions) => {
     if (! context.rigHost) return;
     console.log(`uninstallMiner ${minerName} / ${minerAlias}`);
 
@@ -96,6 +114,7 @@ export const uninstallMiner = (context: GlobalContextType, minerName: string, mi
     const onSuccess = (minerName: string, minerAlias: string | undefined, result: any) => {};
     const onFail = (minerName: string, minerAlias: string | undefined, err: any) => {};
 
+    const confirmed = options?.confirmed || false;
     let error: string | null = null;
 
     if (! minerName) {
@@ -110,48 +129,55 @@ export const uninstallMiner = (context: GlobalContextType, minerName: string, mi
     //const minerFullName = `${minerName}-${minerAlias}`;
     const minerFullTitle = (minerName === minerAlias || ! minerAlias) ? minerName : `${minerName} (${minerAlias}))`;
 
-    alertify.confirm("<b>Miner uninstallation - confirmation</b>", `Do you want to uninstall the miner '<b>${minerFullTitle}</b>' ?`,
-        function(){
-            alertify.success(`Starting miner ${minerFullTitle} uninstallation...`);
+    function _uninstallMiner() {
+        alertify.success(`Starting miner ${minerFullTitle} uninstallation...`);
 
-            if (typeof onStart === 'function') {
-                onStart(minerName, minerAlias);
-            }
-
-            const data: {[key: string]: any} = {
-                action: 'start',
-                miner: minerName,
-                minerAlias,
-            };
-
-            const url = `http://${context.rigHost}/rig/miners/${minerName}/uninstall`;
-
-            fetchHtml(url, { useProxy: true, method: 'POST', body: JSON.stringify(data) })
-                .then(({data, headers, status}) => {
-                    if (data && data.startsWith('OK:')) {
-                        if (typeof onSuccess === 'function') {
-                            onSuccess(minerName, minerAlias, data);
-                        }
-                        alertify.success(`Miner ${minerFullTitle} uninstallation started`);
-
-                    } else {
-                        if (typeof onFail === 'function') {
-                            onFail(minerName, minerAlias, { message: data });
-                        }
-                        alertify.error(`Miner ${minerFullTitle} uninstallation cannot be started. ${data}`);
-                    }
-
-                })
-                .catch((err: any) => {
-                    if (typeof onFail === 'function') {
-                        onFail(minerName, minerAlias, err);
-                    }
-
-                    alertify.error(`Miner ${minerFullTitle} uninstallation cannot be started. ${err.message}`);
-                });
-        },
-        function(){
-            //alertify.error('Cancel');
+        if (typeof onStart === 'function') {
+            onStart(minerName, minerAlias);
         }
-    );
+
+        const data: {[key: string]: any} = {
+            action: 'start',
+            miner: minerName,
+            minerAlias,
+        };
+
+        const url = `http://${context.rigHost}/rig/miners/${minerName}/uninstall`;
+
+        fetchHtml(url, { useProxy: true, method: 'POST', body: JSON.stringify(data) })
+            .then(({data, headers, status}) => {
+                if (data && data.startsWith('OK:')) {
+                    if (typeof onSuccess === 'function') {
+                        onSuccess(minerName, minerAlias, data);
+                    }
+
+                    alertify.success(`Miner ${minerFullTitle} uninstallation started`);
+
+                } else {
+                    if (typeof onFail === 'function') {
+                        onFail(minerName, minerAlias, { message: data });
+                    }
+
+                    alertify.error(`Miner ${minerFullTitle} uninstallation cannot be started. ${data}`);
+                }
+
+            })
+            .catch((err: any) => {
+                if (typeof onFail === 'function') {
+                    onFail(minerName, minerAlias, err);
+                }
+
+                alertify.error(`Miner ${minerFullTitle} uninstallation cannot be started. ${err.message}`);
+            });
+    }
+
+    if (confirmed) {
+        _uninstallMiner();
+
+    } else {
+        alertify.confirm("<b>Miner uninstallation - confirmation</b>", `Do you want to uninstall the miner '<b>${minerFullTitle}</b>' ?`,
+            _uninstallMiner,
+            () => { /* alertify.error('Cancel'); */ }
+        );
+    }
 };
